% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUTMzone.R
\name{getUTMzone}
\alias{getUTMzone}
\title{Get UTM Zone from an sf Object}
\usage{
getUTMzone(sf_object, epsg_code = FALSE)
}
\arguments{
\item{sf_object}{An object of class \code{sf}.}

\item{epsg_code}{Boolean (FALSE/TRUE). If TRUE, the function will return the EPSG code of the UTM zone. Default is FALSE.}
}
\value{
An integer value representing the UTM zone.
}
\description{
This function calculates the UTM zone for a given sf object by computing its bounding box,
determining the centroid of that box, transforming the centroid to EPSG:4326 (WGS84),
and then calculating the UTM zone based on the centroid's longitude.
}
\details{
The UTM zone is computed using the formula:
\deqn{floor((longitude + 180) / 6) + 1,}
where \code{longitude} is extracted from the centroid of the bounding box after transforming
to EPSG:4326. This function works regardless of the original coordinate reference system of
the sf object.
}
\examples{
\dontrun{
  library(sf)
  my_sf <- st_read("path/to/your/shapefile.shp")
  utm_zone <- getUTMzone(my_sf)
  print(utm_zone)
}

}
