% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationships_to_index.R
\name{relationships_to_index}
\alias{relationships_to_index}
\title{Relationships table to Upstream reach ID Idex Table}
\usage{
relationships_to_index(rel)
}
\arguments{
\item{rel}{A \code{data.table} or \code{data.frame} with two integer columns:
\describe{
\item{fromedge}{The upstream reach ID.}
\item{toedge}{The downstream reach ID.}
}}
}
\value{
A \code{data.table} with integer columns:
\describe{
\item{rid}{Downstream reach ID.}
\item{usrid}{Upstream reach ID (includes the reach itself).}
}
}
\description{
Given a table of directed stream‐network edges (\code{fromedge} → \code{toedge}), this function
summarizes each downstream reach (\code{rid}) every upstream reach ID (\code{usrid}), including itself.
}
\details{
This routine iteratively expands the set of upstream relationships via fast keyed joins
in \strong{data.table}.  It first seeds each node as upstream of itself, adds all direct
parents, then “peels back” generation by generation until no new pairs appear.
}
\examples{
library(data.table)
rel <- data.table(fromedge = c(2, 3, 4, 9, 14, 15), toedge = c(1, 2, 2, 10, 10, 14))
# For rid = 1, usrid will be 1,2,3,4; for rid = 2, usrid = 2,3,4; etc.
relationships_to_index(rel)

}
